Refactor Instrument classes
