+++ WebApp/src/uk/ac/exeter/QuinCe/data/Instrument/Instrument.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.Instrument;
 
 import java.io.Serializable;
 import java.text.ParseException;
@@ -7,6 +7,9 @@ import java.util.Calendar;
 import java.util.List;
 import java.util.TreeSet;
 
+import uk.ac.exeter.QuinCe.data.DateTimeParseException;
+import uk.ac.exeter.QuinCe.data.RawDataFileException;
+import uk.ac.exeter.QuinCe.data.RunType;
 import uk.ac.exeter.QuinCe.database.DatabaseUtils;
 import uk.ac.exeter.QuinCe.utils.DateTimeUtils;
 import uk.ac.exeter.QuinCe.utils.MissingParamException;
