+++ WebApp/src/uk/ac/exeter/QuinCe/data/NoDataQCRecord.java
@@ -9,6 +9,7 @@ import uk.ac.exeter.QCRoutines.data.DataRecordException;
 import uk.ac.exeter.QCRoutines.messages.Flag;
 import uk.ac.exeter.QCRoutines.messages.Message;
 import uk.ac.exeter.QCRoutines.messages.MessageException;
+import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.database.Calculation.RawDataDB;
 
 /**
