+++ WebApp/src/uk/ac/exeter/QuinCe/data/RawDataFile.java
@@ -11,6 +11,8 @@ import java.util.Arrays;
 import java.util.Calendar;
 import java.util.List;
 
+import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
+import uk.ac.exeter.QuinCe.data.Instrument.InstrumentException;
 import uk.ac.exeter.QuinCe.utils.DateTimeUtils;
 
 /**
