+++ WebApp/src/uk/ac/exeter/QuinCe/web/Instrument/InstrumentListBean.java
@@ -2,8 +2,8 @@ package uk.ac.exeter.QuinCe.web.Instrument;
 
 import java.util.List;
 
-import uk.ac.exeter.QuinCe.data.InstrumentStub;
-import uk.ac.exeter.QuinCe.database.Instrument.InstrumentDB;
+import uk.ac.exeter.QuinCe.data.Instrument.InstrumentDB;
+import uk.ac.exeter.QuinCe.data.Instrument.InstrumentStub;
 import uk.ac.exeter.QuinCe.web.BaseManagedBean;
 import uk.ac.exeter.QuinCe.web.system.ServletUtils;
